refactor: move recordings-tree path helpers out of WatchRecordsLoop
