+++ modules/core/src/test/kotlin/ru/zinin/frigate/analyzer/core/task/WatchRecordsLoopTest.kt
@@ -7,10 +7,7 @@ import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertNotEquals
-import org.junit.jupiter.api.Assertions.assertNull
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import ru.zinin.frigate.analyzer.core.config.properties.RecordsWatcherProperties
 import ru.zinin.frigate.analyzer.model.dto.RecordingFileDto
@@ -32,8 +29,6 @@ import java.time.ZoneOffset
 import java.util.UUID
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
-
-private val ROOT = Path.of("/mnt/data/frigate/recordings")
 
 class WatchRecordsLoopTest {
     private val recordingEntityHelper = mockk<RecordingEntityHelper>()
@@ -221,85 +216,4 @@ class WatchRecordsLoopTest {
                 tmpDir.toFile().deleteRecursively()
             }
         }
-
-    // --- Pure-function tests migrated from WatchRecordsTaskTest ---
-
-    @Test
-    fun `extractDateFromPath returns date for date directory`() {
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-15")
-        assertEquals(LocalDate.of(2026, 2, 15), extractDateFromPath(path, ROOT))
-    }
-
-    @Test
-    fun `extractDateFromPath returns date for hour subdirectory`() {
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-15/09")
-        assertEquals(LocalDate.of(2026, 2, 15), extractDateFromPath(path, ROOT))
-    }
-
-    @Test
-    fun `extractDateFromPath returns date for camera subdirectory`() {
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-15/09/cam1")
-        assertEquals(LocalDate.of(2026, 2, 15), extractDateFromPath(path, ROOT))
-    }
-
-    @Test
-    fun `extractDateFromPath returns null for root recordings directory`() {
-        val path = Path.of("/mnt/data/frigate/recordings")
-        assertNull(extractDateFromPath(path, ROOT))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns true for today`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-15")
-        assertTrue(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns true for yesterday within 1 day period`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-14")
-        assertTrue(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns false for old date`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings/2026-01-01")
-        assertFalse(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns true for root directory without date`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings")
-        assertTrue(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns true for exact cutoff date`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-14")
-        assertTrue(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `isWithinWatchPeriod returns false for one day before cutoff`() {
-        val clock = Clock.fixed(Instant.parse("2026-02-15T12:00:00Z"), ZoneOffset.UTC)
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-13")
-        assertFalse(isWithinWatchPeriod(path, ROOT, Duration.ofDays(1), clock))
-    }
-
-    @Test
-    fun `extractDateFromPath returns null for invalid date like 2026-02-30`() {
-        val path = Path.of("/mnt/data/frigate/recordings/2026-02-30")
-        assertNull(extractDateFromPath(path, ROOT))
-    }
-
-    @Test
-    fun `extractDateFromPath ignores date-like segments in root path`() {
-        val rootWithDate = Path.of("/data/2024-01-15/frigate/recordings")
-        val path = Path.of("/data/2024-01-15/frigate/recordings/2026-02-15/09/cam1")
-        assertEquals(LocalDate.of(2026, 2, 15), extractDateFromPath(path, rootWithDate))
-    }
 }
